requirements.txt
streamlit
opencv-python-headless
pytesseract
Pillow
pdf2image
numpy
pandas
